+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Note: Numbers like (#123) point to closed Pull Requests on the fractal-vizarr-vi
 # 0.4.2
 
 * Added workaround to fix labels bug in most recent Chrome versions (\#68);
+* Drop `TestingBasicAuthAuthorizer` and `NoneAuthorizer` (\#75);
 
 # 0.4.0